+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.2.0] - 2023-12-22
+
+- Add `{% partial NAME %}` tag to define a partial inside a template
+- Add `render_partial` utility function to render a partial
+- Add `partial=XXX` argument to components to render only a partial
+- Improve docs
+
 ## [0.1.1] - 2023-12-18
 
 - Bugfixes